Improvements and fixes:
[FIX] cfdiv32mx: Added a test when invoice is asked with an erroneous document.
There we want ensure the document attribute is filled to check the generated
document, then behavior was not correct because debug_mode was not being setted
correctly in the object it was being wired to False (stupid error).
[IMP] cfdilib: When debug mode is enabled, it do not make too much
sense do something different from a valid document, simply added a validation
to behave only different if document is invalid AND debug_mode=True.
[IMP] cfdilib: Minimal PEP8 fix.
[FIX] .gitignore: Working with pycharm .idea folder should be ignored to not
commit it by mistake, this folder should be added in a different commit to have
the configuration correctly setted.
[FIX] various: Some typo on doc string.